lazy-loading (#59)
* lazy-loading

* Fixing Linter Errors

* Removing need to dimportrequire admin within smaller scope

* Separating analytics endpoint to reduce cold start

* Modifying Loggers
