--- functions/utils/admin.js
@@ -1,6 +1,0 @@
-const admin = require("firebase-admin");
-admin.initializeApp();
-db = admin.firestore();
-storage = admin.storage();
-
-module.exports = { db, admin, storage };
